First version of C++ code generator
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@857 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/serial/datatool/parser.hpp
@@ -34,7 +34,7 @@ public:
     void TypeList(list<string>& ids);
     AutoPtr<ASNValue> Value(void);
     AutoPtr<ASNValue> x_Value(void);
-    long Number(void);
+    int Number(void);
     string String(void);
     string Identifier(void);
     string TypeReference(void);
